Word counter
This program was designed to read words from a file and count them, with the option of displaying the longest word, shortest word or both; at the end the program will display a the amount of words in the document as well. If the document doesn't exist the program will read input from the command line with the option of displaying the longest word, shortest word or both, and display the total amount of words.