Binary search tree structure
This program was designed to read sentences from a file and compare them by their asccii value, with the option of case sensitive analysis; at the end the program the sentences will be printed into a output file or console in order. If the document doesn't exist the program will read input from the command line.